backend/requirements.txt
fastapi
uvicorn
pydantic
python-multipart
beautifulsoup4
requests
python-dotenv
Pillow
yt-dlp
